Bulk changes. bad me.
So I finally figured out why mongo wasn't working. I was including it in
client-side code which was all fucked up. Obviously I don't need it in
the client since the server will just get the data from mongo and send
it to the client, which means the client doesn't need to know where it
even comes from.